# dfu-service/src/main/java/cn/gatesma/desirefu/controller/api/generate/CollegeApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.8).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package cn.gatesma.desirefu.controller.api.generate;

import cn.gatesma.desirefu.domain.api.generate.AddCollegeRequest;
import cn.gatesma.desirefu.domain.api.generate.AddCollegeRet;
import cn.gatesma.desirefu.domain.api.generate.GetCollegeRequest;
import cn.gatesma.desirefu.domain.api.generate.GetCollegeRet;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@Api(value = "College", tags = {"高校"}, description = "the College API")
public interface CollegeApi {

    @ApiOperation(value = "创建学校", nickname = "addCollege", notes = "创建学校，返回\"学校ID\"", response = AddCollegeRet.class, tags={ "高校", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = AddCollegeRet.class),
        @ApiResponse(code = 400, message = "Invalid RequestBody supplied"),
        @ApiResponse(code = 404, message = "RequestBody not found") })
    @RequestMapping(value = "/college/add",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<AddCollegeRet> addCollege(@ApiParam(value = "创建学校" ,required=true )  @Valid @RequestBody AddCollegeRequest body);

    @ApiOperation(value = "查询高校", nickname = "getCollege", notes = "查询高校", response = GetCollegeRet.class, tags={ "高校", })
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "OK", response = GetCollegeRet.class),
            @ApiResponse(code = 400, message = "Invalid RequestBody supplied"),
            @ApiResponse(code = 404, message = "RequestBody not found") })
    @RequestMapping(value = "/college/get",
            produces = { "application/json" },
            consumes = { "application/json" },
            method = RequestMethod.POST)
    ResponseEntity<GetCollegeRet> getCollege(@ApiParam(value = "查询高校" ,required=true )  @Valid @RequestBody GetCollegeRequest body);


}