# dfu-service/src/main/resources/swagger.yaml
...
swagger: "2.0"
info:
  description: DesireFU后端接口文档。
  version: 1.0.0
  title: DFU_API_DOC
host: gatesma.cn
basePath: /desirefu/v1
...
schemes:
  - http
paths:
...
  /college/add:
    post:
      tags:
        - College
      summary: 创建学校
      description: 创建学校，返回"学校ID"
      operationId: addCollege
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: 创建学校
          required: true
          schema:
            $ref: '#/definitions/AddCollegeRequest'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/AddCollegeRet'
        400:
          description: Invalid RequestBody supplied
        404:
          description: RequestBody not found
  /college/get:
    post:
      tags:
        - College
      summary: 查询高校
      description: 查询高校
      operationId: getCollege
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: 查询高校
          required: true
          schema:
            $ref: '#/definitions/GetCollegeRequest'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/GetCollegeRet'
        400:
          description: Invalid RequestBody supplied
        404:
          description: RequestBody not found

  # department
...
definitions:
...
  ReturnCode:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
        example: 0
      message:
        type: string
        example: OK
...
  AddCollegeRet:
    allOf:
      - $ref: '#/definitions/ReturnCode'
      - properties:
          data:
            $ref: '#/definitions/AddCollegeData'
  AddCollegeData:
    properties:
      college_id:
        type: integer
        format: int32
        example: 2211
        description: 学校ID
  AddCollegeRequest:
    required:
      - name
      - ministry
      - identification
    properties:
      name:
        type: string
        example: 四川大学
        description: 学校名称
      ministry:
        type: string
        example: 教育部
        description: 主管部门
      identification:
        type: string
        example: 4151010610
        description: 学校标识码
      location:
        type: string
        example: 成都
        description: 所在地
      level:
        type: string
        example: 本科
        description: 办学层次
      memo:
        type: string
        example: 备注
        description: 备注
...
  GetCollegeRequest:
    properties:
      name:
        type: string
        example: 四川大学
        description: 高校名称
      ministry:
        type: string
        example: 教育部
        description: 主管部门
      identification:
        type: string
        example: 4151010610
        description: 唯一标识码
  GetCollegeRet:
    allOf:
      - $ref: '#/definitions/ReturnCode'
      - properties:
          data:
            type: array
            items:
              $ref: '#/definitions/College'
  College:
    description: 高校
    properties:
      college_id:
        type: integer
        format: int32
        example: 2211
        description: 学校ID
      name:
        type: string
        example: 四川大学
        description: 学校名称
      ministry:
        type: string
        example: 教育部
        description: 主管部门
      identification:
        type: string
        example: 4151010610
        description: 学校标识码
      location:
        type: string
        example: 成都
        description: 所在地
      level:
        type: string
        example: 本科
        description: 办学层次
      memo:
        type: string
        example: 备注
        description: 备注
      created_time:
        type: string
        example: "2020-01-01 00:00:00"
        description: 创建时间，格式：yyyy-MM-dd HH:mm:ss
      last_modified_time:
        type: string
        example: "2000-01-01 00:00:00"
        description: 最后修改时间，格式：yyyy-MM-dd HH:mm:ss
...